Add files via upload
Due to file constraints, this is a condensed demonstration of the program via a ipynb file taken from the Jupyter Notebook file. There is a user guide for to demonstrate the usage of the file, but due to Github constraints, the image files are not included. 

An included copy of the full breakdown of the project has been included (Assessment 2.docx)

The program was written in Python via Jupyter Notebooks